pcap: split on windows/non-windows functionality
Create pcap_windows and pcap_unix for windows and non-windows
functionality, respectively. libpcap does not support
pcap_get_selectable_fd on Windows so we cannot use it. In the best case,
this issue does not exist on Windows.
